requirements.txt
pandas
pickle
streamlit
numpy
matplotlib
xgboost
sklearn.model_selection
